+++ Changelog.md
@@ -11,6 +11,7 @@
 - Fix: Inventory scanner crash
 - Fix: Placing a Secure Redstone Interface as a receiver does not correctly update itself and surrounding blocks
 - Fix: Blocks projected by a Projector are reset to their default state when opening the Projector's screen
+- Fix: The owner of the topmost block of the cage spawned by a Cage Trap is not set correctly 
 
 --------------------------Changelog for v1.10 of SecurityCraft--------------------------